+++ relay-event-schema/src/protocol/span.rs
@@ -940,6 +940,11 @@ pub struct SpanData {
     #[metastructure(field = "lcp.url")]
     pub lcp_url: Annotated<String>,
 
+    // The span's name, a brief, human-readable, low cardinality description of operation
+    // represented by the span (as per OpenTelemetry/Sentry's Span V2 schema).
+    #[metastructure(field = "sentry.name")]
+    pub span_name: Annotated<String>,
+
     /// Other fields in `span.data`.
     #[metastructure(
         additional_properties,
@@ -1421,7 +1426,7 @@ mod tests {
             .unwrap()
             .into_value()
             .unwrap();
-        insta::assert_debug_snapshot!(data, @r#"
+        insta::assert_debug_snapshot!(data, @r###"
         SpanData {
             app_start_type: ~,
             gen_ai_request_max_tokens: ~,
@@ -1546,6 +1551,7 @@ mod tests {
             lcp_size: ~,
             lcp_id: ~,
             lcp_url: ~,
+            span_name: ~,
             other: {
                 "bar": String(
                     "3",
@@ -1555,7 +1561,7 @@ mod tests {
                 ),
             },
         }
-        "#);
+        "###);
 
         assert_eq!(data.get_value("foo"), Some(Val::U64(2)));
         assert_eq!(data.get_value("bar"), Some(Val::String("3")));